Разделяет скрипт на модули
