+++ js/constants.js
@@ -1,0 +1,36 @@
+export const AD_NAMES = [
+  'ОливаАрт',
+  'Эсмеральда',
+  'Пушка Инн',
+  'Охтинская',
+  'Окунёвая',
+  'ИжПарк',
+  'Ривьера',
+  'Сальвадор',
+  'SkyPort',
+  'СКИ ИНН',
+];
+export const AD_DESCRIPTIONS = [
+  'Новый современный отель в Алуште',
+  'Отель все включено «Эмеральд» готов предложить своим гостям отличный, комфортный отдых в Витязево',
+  'Отель «Пушка ИНН» возрождает традиции Петербургского гостеприимства',
+  'Приглашаем вас в отель Охтинская – на берегу Невы вы отлично проведете время',
+  'База Отдыха Окунёвая расположилась в Ленинградской области, на берегу живописной Дубковой бухты Финского залива',
+  '«Парк-Отель» - мини-отель категории 5 звёзд с 11 просторными номерами',
+  '«Ривьера» - четырехзвездочный отель в центре Казани на набережной реки Казанка',
+  'Отличный отель второй пляжной линии на Толстом мысе',
+  'Отель SKYPORT 4* расположен в 300 метрах от Международного аэропорта «Толмачево»',
+  'Самое сердце горного курорта «Роза Хутор» на высоте 1170 метров',
+];
+
+export const AD_TYPES = ['palace', 'flat', 'house', 'bungalow', 'hotel'];
+export const AD_CHECKIN_TIMES = ['12:00', '13:00', '14:00'];
+export const AD_CHECKOUT_TIMES = ['12:00', '13:00', '14:00'];
+export const AD_FEATURES = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'];
+export const AD_PHOTOS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+];
+
+export const AD_COUNT = 10;
